finished loading in normalized values.
will now focus on storing distance of closest points in linkedhashmap.

+++ SURI CSV Reader/src/Pair.java
@@ -1,25 +1,25 @@
 import javafx.beans.binding.DoubleExpression;
 
 public class Pair{
-    Double p1, p2;
-        Pair(Double p1, Double p2) {
+    long p1, p2;
+        Pair(long p1, long p2) {
             this.p1 = p1;
             this.p2 = p2;
         }
 
-    public Double getP1() {
+    public long getP1() {
         return p1;
     }
 
-    public void setP1(Double p1) {
+    public void setP1(long p1) {
         this.p1 = p1;
     }
 
-    public void setP2(Double p2) {
+    public void setP2(long p2) {
         this.p2 = p2;
     }
 
-    public Double getP2() {
+    public long getP2() {
         return p2;
     }
